ai/requirement.txt
langchain
langchain-openai
python-dotenv
langchain-google-firestore
langchain-google-genai
langchain-chroma
pypdf